Export functionality to talons and use within checkout components
