+++ packages/venia-ui/lib/components/Checkout/Receipt/receipt.js
@@ -1,45 +1,24 @@
-import React, { Fragment, useCallback, useEffect, useRef } from 'react';
+import React, { Fragment } from 'react';
 import { compose } from 'redux';
 import { func, shape, string } from 'prop-types';
 
 import { mergeClasses } from '../../../classify';
 import Button from '../../Button';
 import defaultClasses from './receipt.css';
-import { useCheckoutContext } from '@magento/peregrine/lib/context/checkout';
-import { useUserContext } from '@magento/peregrine/lib/context/user';
-import { useAppContext } from '@magento/peregrine/lib/context/app';
 import { withRouter } from '@magento/venia-drivers';
+import { useReceipt } from '@magento/peregrine/lib/talons/Checkout/Receipt/useReceipt';
 
 /**
  * A component that displays some basic information about an order and has
  * a call to action for viewing order details and creating an account.
  */
 const Receipt = props => {
-    const { history, onClose } = props;
-
-    const [{ drawer }] = useAppContext();
-    const [, { createAccount, resetReceipt }] = useCheckoutContext();
-    const [{ isSignedIn }] = useUserContext();
+    const {
+        handleCreateAccount,
+        handleViewOrderDetails,
+        isSignedIn
+    } = useReceipt(props);
     const classes = mergeClasses(defaultClasses, props.classes);
-
-    // When the drawer is closed reset the state of the receipt. We use a ref
-    // because drawer can change if the mask is clicked. Mask updates drawer.
-    const prevDrawer = useRef(null);
-    useEffect(() => {
-        if (prevDrawer.current === 'cart' && drawer !== 'cart') {
-            resetReceipt();
-            onClose();
-        }
-        prevDrawer.current = drawer;
-    }, [drawer, onClose, resetReceipt]);
-
-    const handleCreateAccount = useCallback(() => {
-        createAccount(history);
-    }, [createAccount, history]);
-
-    const handleViewOrderDetails = useCallback(() => {
-        // TODO: Implement/connect/redirect to order details page.
-    }, []);
 
     return (
         <div className={classes.root}>
