+++ packages/venia-ui/lib/components/Checkout/__tests__/addressForm.spec.js
@@ -7,8 +7,8 @@ jest.mock('../../../classify');
 const mockCancel = jest.fn();
 const mockSubmit = jest.fn();
 const defaultProps = {
-    cancel: mockCancel,
-    submit: mockSubmit
+    onCancel: mockCancel,
+    onSubmit: mockSubmit
 };
 
 beforeEach(() => {
