+++ packages/venia-ui/lib/components/Checkout/addressForm.js
@@ -1,4 +1,4 @@
-import React, { useMemo } from 'react';
+import React from 'react';
 import { Form } from 'informed';
 import { array, bool, func, object, shape, string } from 'prop-types';
 
@@ -14,6 +14,7 @@ import {
 import combine from '../../util/combineValidators';
 import TextInput from '../TextInput';
 import Field from '../Field';
+import { useAddressForm } from '@magento/peregrine/lib/talons/Checkout/useAddressForm';
 
 const fields = [
     'city',
@@ -27,28 +28,27 @@ const fields = [
 ];
 
 const AddressForm = props => {
+    const talonProps = useAddressForm({
+        ...props,
+        fields
+    });
+
     const {
-        cancel,
         countries,
-        initialValues,
-        isSubmitting,
         error,
-        submit
-    } = props;
+        handleCancel,
+        handleSubmit,
+        initialValues,
+        isSubmitting
+    } = talonProps;
 
     const classes = mergeClasses(defaultClasses, props.classes);
-
-    const values = useMemo(
-        () =>
-            fields.reduce((acc, key) => {
-                acc[key] = initialValues[key];
-                return acc;
-            }, {}),
-        [initialValues]
-    );
-
     return (
-        <Form className={classes.root} initialValues={values} onSubmit={submit}>
+        <Form
+            className={classes.root}
+            initialValues={initialValues}
+            onSubmit={handleSubmit}
+        >
             <div className={classes.body}>
                 <h2 className={classes.heading}>Shipping Address</h2>
                 <div className={classes.validationMessage}>{error}</div>
@@ -130,7 +130,7 @@ const AddressForm = props => {
                 </div>
             </div>
             <div className={classes.footer}>
-                <Button onClick={cancel}>Cancel</Button>
+                <Button onClick={handleCancel}>Cancel</Button>
                 <Button type="submit" priority="high" disabled={isSubmitting}>
                     Use Address
                 </Button>
@@ -140,7 +140,7 @@ const AddressForm = props => {
 };
 
 AddressForm.propTypes = {
-    cancel: func.isRequired,
+    onCancel: func.isRequired,
     classes: shape({
         body: string,
         button: string,
@@ -161,7 +161,7 @@ AddressForm.propTypes = {
     error: string,
     initialValues: object,
     isSubmitting: bool,
-    submit: func.isRequired
+    onSubmit: func.isRequired
 };
 
 AddressForm.defaultProps = {
